# src/main/java/ru/bank/cards/entity/TransferStatus.java
package ru.bank.cards.entity;

/**
 * Статус перевода.
 */
public enum TransferStatus {

    COMPLETED("Завершена"),
    FAILED("Отменена");

    final String status;

    TransferStatus(String status) {
        this.status = status;
    }

    @Override
    public String toString() {
        return "Status{" +
                "status='" + status + '\'' +
                '}';
    }
}